Move grbl_cli logic into lib file
